+++ .github/workflows/gdextension-build.yml
@@ -72,7 +72,7 @@ jobs:
 
       - name: Install Linux build dependencies
         if: steps.source.outputs.available == 'true' && matrix.platform == 'linux'
-        run: sudo apt-get update && sudo apt-get install -y build-essential pkg-config libgtk-3-dev libwebkit2gtk-4.1-dev
+        run: sudo apt-get update && sudo apt-get install -y build-essential pkg-config
 
       - name: Compute godot-cpp cache key
         if: steps.source.outputs.available == 'true'
