+++ .github/workflows/package-preview.yml
@@ -50,7 +50,7 @@ jobs:
 
       - name: Install Linux build dependencies
         if: matrix.platform == 'linux'
-        run: sudo apt-get update && sudo apt-get install -y build-essential pkg-config libgtk-3-dev libwebkit2gtk-4.1-dev
+        run: sudo apt-get update && sudo apt-get install -y build-essential pkg-config
 
       - name: Set up Rust
         uses: dtolnay/rust-toolchain@stable
