+++ fennara-cpp/include/fennara/ui/webview_host.hpp
@@ -3,13 +3,17 @@
 #include <godot_cpp/classes/control.hpp>
 #include <godot_cpp/variant/string.hpp>
 
-#include <string>
+#include <memory>
 
 namespace fennara {
 
+namespace webview_backend {
+class NativeWebviewBackend;
+}
+
 class WebviewHost {
 public:
-    WebviewHost() = default;
+    WebviewHost();
     ~WebviewHost();
 
     bool start(godot::Control *owner, const godot::String &url);
@@ -19,20 +23,7 @@ public:
     bool is_started() const;
 
 private:
-    void *webview = nullptr;
-    void *widget = nullptr;
-    void *parent_window = nullptr;
-    godot::String current_url;
-    godot::String helper_state_path;
-    godot::String helper_log_path;
-    bool started = false;
-    int32_t helper_pid = 0;
-    int current_window_id = -1;
-    int last_x = -1;
-    int last_y = -1;
-    int last_width = -1;
-    int last_height = -1;
-    int last_visible_state = -1;
+    std::unique_ptr<webview_backend::NativeWebviewBackend> backend;
 };
 
 } // namespace fennara
