+++ fennara-cpp/src/ui/webview_host.cpp
@@ -1,35 +1,27 @@
 #include "fennara/ui/webview_host.hpp"
 
+#include "webview_backend.hpp"
+
 #include "fennara/logger.hpp"
 
 #include <godot_cpp/classes/display_server.hpp>
-#include <godot_cpp/classes/dir_access.hpp>
-#include <godot_cpp/classes/file_access.hpp>
 #include <godot_cpp/classes/os.hpp>
-#include <godot_cpp/classes/project_settings.hpp>
-#include <godot_cpp/classes/window.hpp>
 #include <godot_cpp/variant/utility_functions.hpp>
-
-#ifdef _WIN32
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
-#include <webview/webview.h>
-#endif
-
-#include <string>
 
 namespace fennara {
 
-#ifdef __APPLE__
-namespace mac_webview {
-bool start(void **webview, void **parent_window, godot::Control *owner, const godot::String &url);
-void resize_to(void *webview, void **parent_window, godot::Control *owner);
-void set_visible(void *webview, bool visible);
-void stop(void **webview, void **parent_window);
-} // namespace mac_webview
-#endif
-
 namespace {
+
+bool editor_is_headless() {
+    godot::DisplayServer *display = godot::DisplayServer::get_singleton();
+    godot::OS *os = godot::OS::get_singleton();
+    return (os != nullptr && os->has_feature("headless")) ||
+           (display != nullptr && display->get_name().to_lower() == "headless");
+}
+
+} // namespace
+
+namespace webview_backend {
 
 void output_log(const godot::String &message) {
     FLOG_UI(message);
@@ -41,386 +33,54 @@ void output_error(const godot::String &message) {
     godot::UtilityFunctions::push_error(godot::String("[Fennara] ") + message);
 }
 
-struct WebviewGeometry {
-    bool visible = false;
-    int x = 0;
-    int y = 0;
-    int width = 0;
-    int height = 0;
-};
+} // namespace webview_backend
 
-int owner_window_id(godot::Control *owner) {
-    if (owner == nullptr) {
-        return 0;
-    }
-    godot::Window *window = owner->get_window();
-    if (window == nullptr) {
-        return 0;
-    }
-    return window->get_window_id();
+WebviewHost::WebviewHost() :
+        backend(webview_backend::create_backend()) {
 }
-
-WebviewGeometry compute_webview_geometry(godot::Control *owner) {
-    WebviewGeometry geometry;
-    if (owner == nullptr || !owner->is_visible_in_tree()) {
-        return geometry;
-    }
-
-    godot::Vector2 size = owner->get_size();
-    if (size.x <= 0 || size.y <= 0) {
-        return geometry;
-    }
-
-    godot::Vector2 screen_position = owner->get_screen_position();
-    geometry.visible = true;
-    geometry.x = static_cast<int>(screen_position.x);
-    geometry.y = static_cast<int>(screen_position.y);
-    geometry.width = static_cast<int>(size.x);
-    geometry.height = static_cast<int>(size.y);
-    return geometry;
-}
-
-bool editor_is_headless() {
-    godot::DisplayServer *display = godot::DisplayServer::get_singleton();
-    godot::OS *os = godot::OS::get_singleton();
-    return (os != nullptr && os->has_feature("headless")) ||
-           (display != nullptr && display->get_name().to_lower() == "headless");
-}
-
-#ifdef __linux__
-godot::String globalize_project_path(const godot::String &path) {
-    godot::ProjectSettings *settings = godot::ProjectSettings::get_singleton();
-    return settings == nullptr ? path : settings->globalize_path(path);
-}
-
-godot::String linux_helper_binary_path() {
-    return globalize_project_path("res://addons/fennara/bin/fennara-chat-webview");
-}
-
-godot::String linux_helper_state_path() {
-    return globalize_project_path("user://.fennara/webview-helper-state.txt");
-}
-
-godot::String linux_helper_log_path() {
-    return globalize_project_path("user://.fennara/webview-helper.log");
-}
-
-void ensure_parent_dir(const godot::String &path) {
-    const godot::String base_dir = path.get_base_dir();
-    if (!base_dir.is_empty()) {
-        godot::DirAccess::make_dir_recursive_absolute(base_dir);
-    }
-}
-
-bool write_linux_helper_state(const godot::String &path, const WebviewGeometry &geometry) {
-    if (path.is_empty()) {
-        return false;
-    }
-    ensure_parent_dir(path);
-
-    godot::Ref<godot::FileAccess> file =
-        godot::FileAccess::open(path, godot::FileAccess::WRITE);
-    if (file.is_null()) {
-        return false;
-    }
-
-    file->store_string(godot::String("visible=") + (geometry.visible ? "1" : "0") + "\n");
-    file->store_string("x=" + godot::String::num_int64(geometry.x) + "\n");
-    file->store_string("y=" + godot::String::num_int64(geometry.y) + "\n");
-    file->store_string("width=" + godot::String::num_int64(geometry.width) + "\n");
-    file->store_string("height=" + godot::String::num_int64(geometry.height) + "\n");
-    return true;
-}
-#endif
-
-} // namespace
 
 WebviewHost::~WebviewHost() {
     stop();
 }
 
 bool WebviewHost::start(godot::Control *owner, const godot::String &url) {
-    if (started) {
-        output_log("Web chat host already started");
+    if (is_started()) {
+        webview_backend::output_log("Web chat host already started");
         return true;
     }
 
     if (editor_is_headless()) {
-        output_log("Web chat host skipped: headless editor has no native window");
+        webview_backend::output_log("Web chat host skipped: headless editor has no native window");
         return false;
     }
 
-#ifdef _WIN32
-    if (owner == nullptr) {
-        output_error("Web chat host cannot start: owner Control is null");
+    if (backend == nullptr) {
+        webview_backend::output_error("Web chat native webview is not wired for this platform build yet");
         return false;
     }
-
-    godot::DisplayServer *display = godot::DisplayServer::get_singleton();
-    if (display == nullptr) {
-        output_error("Web chat host cannot start: DisplayServer is unavailable");
-        return false;
-    }
-
-    int window_id = owner_window_id(owner);
-    int64_t native_window = display->window_get_native_handle(
-        godot::DisplayServer::WINDOW_HANDLE,
-        window_id);
-    output_log("Web chat native window id=" + godot::String::num_int64(window_id) +
-               " handle=" + godot::String::num_int64(native_window));
-    if (native_window == 0) {
-        output_error("Web chat host cannot start: Godot native window handle is 0");
-        return false;
-    }
-    current_window_id = window_id;
-    parent_window = reinterpret_cast<void *>(native_window);
-
-    output_log("Web chat creating native webview url=" + url);
-    webview = webview_create(0, parent_window);
-    if (webview == nullptr) {
-        output_error("Web chat host cannot start: webview_create returned null");
-        return false;
-    }
-
-    widget = webview_get_native_handle(
-        static_cast<webview_t>(webview),
-        WEBVIEW_NATIVE_HANDLE_KIND_UI_WIDGET);
-    output_log("Web chat native widget handle=" +
-               godot::String::num_int64(reinterpret_cast<int64_t>(widget)));
-    if (widget == nullptr) {
-        output_error("Web chat host cannot start: native widget handle is null");
-        webview_destroy(static_cast<webview_t>(webview));
-        webview = nullptr;
-        parent_window = nullptr;
-        current_window_id = -1;
-        return false;
-    }
-
-    std::string url_utf8 = url.utf8().get_data();
-    webview_navigate(static_cast<webview_t>(webview), url_utf8.c_str());
-    current_url = url;
-    started = true;
-    resize_to(owner);
-    output_log("Web chat native webview started");
-    return true;
-#elif defined(__APPLE__)
-    if (mac_webview::start(&webview, &parent_window, owner, url)) {
-        current_url = url;
-        started = true;
-        resize_to(owner);
-        output_log("Web chat native macOS webview started");
-        return true;
-    }
-    output_error("Web chat native macOS webview could not start");
-    return false;
-#elif defined(__linux__)
-    current_url = url;
-    godot::OS *os = godot::OS::get_singleton();
-    if (os == nullptr) {
-        output_error("Web chat Linux helper cannot start: OS singleton is unavailable");
-        return false;
-    }
-
-    const godot::String helper_path = linux_helper_binary_path();
-    if (!godot::FileAccess::file_exists(helper_path)) {
-        output_error("Web chat Linux helper is missing: " + helper_path);
-        return false;
-    }
-
-    helper_state_path = linux_helper_state_path();
-    helper_log_path = linux_helper_log_path();
-    WebviewGeometry geometry = compute_webview_geometry(owner);
-    if (!write_linux_helper_state(helper_state_path, geometry)) {
-        output_error("Web chat Linux helper cannot write state file: " + helper_state_path);
-        return false;
-    }
-
-    ensure_parent_dir(helper_log_path);
-    godot::PackedStringArray args;
-    args.append("--url");
-    args.append(url);
-    args.append("--state");
-    args.append(helper_state_path);
-    args.append("--log");
-    args.append(helper_log_path);
-    if (os->get_environment("FENNARA_LINUX_WEBVIEW_DEBUG") == "1") {
-        args.append("--debug");
-    }
-
-    helper_pid = os->create_process(helper_path, args, false);
-    if (helper_pid <= 0) {
-        output_error("Web chat Linux helper failed to start: " + helper_path);
-        return false;
-    }
-
-    started = true;
-    last_x = geometry.x;
-    last_y = geometry.y;
-    last_width = geometry.width;
-    last_height = geometry.height;
-    last_visible_state = geometry.visible ? 1 : 0;
-    output_log("Web chat Linux helper started pid=" + godot::String::num_int64(helper_pid) +
-               " state=" + helper_state_path +
-               " log=" + helper_log_path);
-    return true;
-#else
-    (void)owner;
-    (void)url;
-    output_error("Web chat native webview is not wired for this platform build yet");
-    return false;
-#endif
+    return backend->start(owner, url);
 }
 
 void WebviewHost::resize_to(godot::Control *owner) {
-    if (!started || owner == nullptr) {
-        return;
+    if (backend != nullptr) {
+        backend->resize_to(owner);
     }
-
-#ifdef _WIN32
-    if (widget == nullptr) {
-        output_error("Web chat resize skipped: native widget handle is null");
-        return;
-    }
-
-    WebviewGeometry geometry = compute_webview_geometry(owner);
-    HWND hwnd = reinterpret_cast<HWND>(widget);
-    if (!geometry.visible) {
-        ShowWindow(hwnd, SW_HIDE);
-        return;
-    }
-
-    int window_id = owner_window_id(owner);
-    if (window_id != current_window_id) {
-        output_log("Web chat recreating native webview for window id=" +
-                   godot::String::num_int64(window_id));
-        godot::String url = current_url;
-        stop();
-        start(owner, url);
-        return;
-    }
-
-    HWND parent_hwnd = reinterpret_cast<HWND>(parent_window);
-    POINT origin{
-        static_cast<LONG>(geometry.x),
-        static_cast<LONG>(geometry.y),
-    };
-    if (parent_hwnd != nullptr) {
-        ScreenToClient(parent_hwnd, &origin);
-    }
-    int x = static_cast<int>(origin.x);
-    int y = static_cast<int>(origin.y);
-
-    SetWindowPos(
-        hwnd,
-        nullptr,
-        x,
-        y,
-        geometry.width,
-        geometry.height,
-        SWP_NOACTIVATE | SWP_NOZORDER);
-    ShowWindow(hwnd, SW_SHOW);
-
-    if (x != last_x || y != last_y || geometry.width != last_width ||
-        geometry.height != last_height) {
-        last_x = x;
-        last_y = y;
-        last_width = geometry.width;
-        last_height = geometry.height;
-        output_log("Web chat geometry x=" + godot::String::num_int64(x) +
-                   " y=" + godot::String::num_int64(y) +
-                   " w=" + godot::String::num_int64(geometry.width) +
-                   " h=" + godot::String::num_int64(geometry.height));
-    }
-#else
-#ifdef __APPLE__
-    mac_webview::resize_to(webview, &parent_window, owner);
-#elif defined(__linux__)
-    WebviewGeometry geometry = compute_webview_geometry(owner);
-    if (geometry.x != last_x || geometry.y != last_y || geometry.width != last_width ||
-        geometry.height != last_height || (geometry.visible ? 1 : 0) != last_visible_state) {
-        if (write_linux_helper_state(helper_state_path, geometry)) {
-            last_x = geometry.x;
-            last_y = geometry.y;
-            last_width = geometry.width;
-            last_height = geometry.height;
-            last_visible_state = geometry.visible ? 1 : 0;
-            output_log("Web chat Linux helper state x=" + godot::String::num_int64(geometry.x) +
-                       " y=" + godot::String::num_int64(geometry.y) +
-                       " w=" + godot::String::num_int64(geometry.width) +
-                       " h=" + godot::String::num_int64(geometry.height) +
-                       " visible=" + godot::String::num_int64(last_visible_state));
-        }
-    }
-#else
-    (void)owner;
-#endif
-#endif
 }
 
 void WebviewHost::set_visible(bool visible) {
-#ifdef _WIN32
-    if (!started || widget == nullptr) {
-        return;
+    if (backend != nullptr) {
+        backend->set_visible(visible);
     }
-    ShowWindow(reinterpret_cast<HWND>(widget), visible ? SW_SHOW : SW_HIDE);
-#elif defined(__APPLE__)
-    mac_webview::set_visible(webview, visible);
-#elif defined(__linux__)
-    WebviewGeometry geometry;
-    geometry.visible = visible;
-    geometry.x = last_x < 0 ? 0 : last_x;
-    geometry.y = last_y < 0 ? 0 : last_y;
-    geometry.width = last_width < 1 ? 420 : last_width;
-    geometry.height = last_height < 1 ? 640 : last_height;
-    if (write_linux_helper_state(helper_state_path, geometry)) {
-        last_visible_state = visible ? 1 : 0;
-        output_log("Web chat Linux helper visibility=" +
-                   godot::String::num_int64(last_visible_state));
-    }
-#else
-    (void)visible;
-#endif
 }
 
 void WebviewHost::stop() {
-    if (!started) {
-        return;
+    if (backend != nullptr) {
+        backend->stop();
     }
-
-#ifdef _WIN32
-    output_log("Web chat destroying native webview");
-    if (webview != nullptr) {
-        webview_destroy(static_cast<webview_t>(webview));
-    }
-#elif defined(__APPLE__)
-    mac_webview::stop(&webview, &parent_window);
-#elif defined(__linux__)
-    if (helper_pid > 0) {
-        godot::OS *os = godot::OS::get_singleton();
-        if (os != nullptr && os->is_process_running(helper_pid)) {
-            os->kill(helper_pid);
-        }
-    }
-    output_log("Web chat Linux helper stopped pid=" + godot::String::num_int64(helper_pid));
-#endif
-
-    webview = nullptr;
-    widget = nullptr;
-    parent_window = nullptr;
-    helper_state_path = "";
-    helper_log_path = "";
-    started = false;
-    helper_pid = 0;
-    current_window_id = -1;
-    last_x = -1;
-    last_y = -1;
-    last_width = -1;
-    last_height = -1;
-    last_visible_state = -1;
 }
 
 bool WebviewHost::is_started() const {
-    return started;
+    return backend != nullptr && backend->is_started();
 }
 
 } // namespace fennara
