+++ scripts/package-preview.mjs
@@ -104,10 +104,7 @@ function isAddonBinary(relative) {
     return relative === "bin/libfennara.windows.editor.x86_64.dll";
   }
   if (platform === "linux") {
-    return (
-      relative === "bin/libfennara.linux.editor.x86_64.so" ||
-      relative === "bin/fennara-chat-webview"
-    );
+    return relative === "bin/libfennara.linux.editor.x86_64.so";
   }
   if (platform === "macos") {
     return (
